requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
Django==3.2.4
django-cors-headers==4.3.0
djangorestframework==3.14.0
eval_type_backport==0.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jsonpath-python==1.0.6
mistralai==1.2.3
mock==5.1.0
mypy-extensions==1.0.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2025.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing-inspect==0.9.0
typing_extensions==4.12.2
